problems arose with doubling all precisions. for now, avoiding compiler warnings by making precision conversions explicit

+++ source/mocassin.f90
@@ -44,9 +44,9 @@ program MoCaSSiN
 
     type(grid_type) :: grid3D(maxGrids)       ! the 3D Cartesian  grid
 
-    real(kind=dp)            :: test                   ! test
+    real            :: test                   ! test
     integer         :: i, iGrid               ! allocation error status
-    real(kind=dp), dimension(2) :: timing              ! cputimer 
+    real, dimension(2) :: timing              ! cputimer 
     integer         :: nhours, nminutes, nseconds
 
     call cpu_time(timing(1))
